+++ integration-test/src/fir/fir.c
@@ -6,11 +6,10 @@
 //===----------------------------------------------------------------------===//
 
 #include "fir.h"
-#include <cstddef>
 
 int fir(int di[N], int idx[N]) {
   int tmp = 0;
-  for (size_t i = 0; i < N; i++)
+  for (unsigned i = 0; i < N; i++)
     tmp += idx[i] * di[N_DEC - i];
   return tmp;
 }
