+++ integration-test/src/matvec/matvec.c
@@ -5,13 +5,12 @@
 //===----------------------------------------------------------------------===//
 
 #include "matvec.h"
-#include <cstddef>
 
 int matvec(int m[N][N], int v[N], int out[N]) {
   int tmp = 0;
-  for (size_t i = 0; i < N; i++) {
+  for (unsigned i = 0; i < N; i++) {
     tmp = 0;
-    for (size_t j = 0; j < N; j++) {
+    for (unsigned j = 0; j < N; j++) {
       tmp += v[j] * m[i][j];
     }
     out[i] = tmp;
